deleteAllowance to command
