+++ packages/hedera-wallet-snap/packages/snap/src/client/SimpleHederaClientImpl.ts
@@ -94,14 +94,6 @@ export class SimpleHederaClientImpl implements SimpleHederaClient {
     return this.#getAccountBalance(this._client);
   }
 
-  async deleteAllowance(options: {
-    assetType: string;
-    assetId: string;
-    spenderAccountId?: string;
-  }): Promise<TxReceipt> {
-    return this.#deleteAllowance(this._client, options);
-  }
-
   async deleteAccount(options: {
     transferAccountId: string;
   }): Promise<TxReceipt> {
@@ -161,81 +153,6 @@ export class SimpleHederaClientImpl implements SimpleHederaClient {
       )
       .setTransferAccountId(options.transferAccountId)
       .freezeWith(client);
-
-    const txResponse = await transaction.execute(client);
-
-    const receipt = await txResponse.getReceipt(client);
-
-    let newExchangeRate;
-    if (receipt.exchangeRate) {
-      newExchangeRate = {
-        ...receipt.exchangeRate,
-        expirationTime: Utils.timestampToString(
-          receipt.exchangeRate.expirationTime,
-        ),
-      };
-    }
-
-    return {
-      status: receipt.status.toString(),
-      accountId: receipt.accountId ? receipt.accountId.toString() : '',
-      fileId: receipt.fileId ? receipt.fileId : '',
-      contractId: receipt.contractId ? receipt.contractId : '',
-      topicId: receipt.topicId ? receipt.topicId : '',
-      tokenId: receipt.tokenId ? receipt.tokenId : '',
-      scheduleId: receipt.scheduleId ? receipt.scheduleId : '',
-      exchangeRate: newExchangeRate,
-      topicSequenceNumber: receipt.topicSequenceNumber
-        ? String(receipt.topicSequenceNumber)
-        : '',
-      topicRunningHash: CryptoUtils.uint8ArrayToHex(receipt.topicRunningHash),
-      totalSupply: receipt.totalSupply ? String(receipt.totalSupply) : '',
-      scheduledTransactionId: receipt.scheduledTransactionId
-        ? receipt.scheduledTransactionId.toString()
-        : '',
-      serials: JSON.parse(JSON.stringify(receipt.serials)),
-      duplicates: JSON.parse(JSON.stringify(receipt.duplicates)),
-      children: JSON.parse(JSON.stringify(receipt.children)),
-    } as TxReceipt;
-  }
-
-  async #deleteAllowance(
-    client: Client,
-    options: {
-      assetType: string;
-      assetId: string;
-      spenderAccountId?: string;
-    },
-  ): Promise<TxReceipt> {
-    let transaction:
-      | AccountAllowanceApproveTransaction
-      | AccountAllowanceDeleteTransaction;
-
-    if (options.assetType === 'HBAR' || options.assetType === 'TOKEN') {
-      transaction = new AccountAllowanceApproveTransaction();
-      if (options.assetType === 'HBAR') {
-        transaction.approveHbarAllowance(
-          client.operatorAccountId as AccountId,
-          options.spenderAccountId as string,
-          0,
-        );
-      } else {
-        transaction.approveTokenAllowance(
-          options.assetId,
-          client.operatorAccountId as AccountId,
-          options.spenderAccountId as string,
-          0,
-        );
-      }
-    } else {
-      transaction =
-        new AccountAllowanceDeleteTransaction().deleteAllTokenNftAllowances(
-          options.assetId,
-          client.operatorAccountId as AccountId,
-        );
-    }
-
-    transaction.freezeWith(client);
 
     const txResponse = await transaction.execute(client);
